Restructures Maven modules and bumps version to 1.0
