+++ command-bus-cdi/src/main/java/de/triology/cb/cdi/CommandProvider.java
