+++ command-bus-cdi/src/main/resources/META-INF/beans.xml
