+++ command-bus-cdi/src/main/resources/META-INF/services/javax.enterprise.inject.spi.Extension
