+++ command-bus-cdi/src/test/java/de/triology/cb/ByeCommand.java
