+++ command-bus-cdi/src/test/java/de/triology/cb/ByeCommandHandler.java
