+++ command-bus-cdi/src/test/java/de/triology/cb/cdi/CommandBusFactory.java
