+++ command-bus-core/src/main/java/de/triology/cb/Command.java
