+++ command-bus-core/src/main/java/de/triology/cb/CommandHandler.java
