+++ command-bus-core/src/main/java/de/triology/cb/decorator/LoggingCommandBus.java
