+++ command-bus-core/src/test/java/de/triology/cb/HelloCommandHandler.java
