+++ command-bus-core/src/test/resources/logback-test.xml
