+++ command-bus-spring/src/main/java/de/triology/cb/spring/Registry.java
