+++ command-bus-spring/src/test/java/de/triology/cb/HelloCommandHandler.java
@@ -1,0 +1,46 @@
+package de.triology.cb;
+
+import org.springframework.beans.factory.annotation.Autowired;
+
+/**
+ * MIT License
+ *
+ * Copyright (c) 2017 TRIOLOGY GmbH
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+public class HelloCommandHandler implements CommandHandler<String, HelloCommand> {
+
+  private MessageCollector messageCollector;
+
+  @Autowired
+  public HelloCommandHandler(MessageCollector messageCollector) {
+    this.messageCollector = messageCollector;
+  }
+
+  @Override
+  public String handle(HelloCommand command) {
+    String message = "hello " + command.getName();
+    if (messageCollector != null) {
+      messageCollector.add(message);
+    }
+    return message;
+  }
+}
