+++ command-bus-spring/src/test/java/de/triology/cb/spring/SpringCommandBusTest.java
@@ -1,0 +1,60 @@
+/**
+ * MIT License
+ *
+ * Copyright (c) 2017 TRIOLOGY GmbH
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+package de.triology.cb.spring;
+
+import de.triology.cb.CommandHandler;
+import de.triology.cb.HelloCommand;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.junit.MockitoJUnitRunner;
+
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
+
+@RunWith(MockitoJUnitRunner.class)
+public class SpringCommandBusTest {
+
+  @Mock
+  private Registry registry;
+
+  @Mock
+  private CommandHandler<String, HelloCommand> handler;
+
+  @InjectMocks
+  private SpringCommandBus commandBus;
+
+  @Test
+  public void execute() {
+    when(registry.get(HelloCommand.class)).thenReturn(handler);
+
+    HelloCommand command = new HelloCommand("bob");
+    commandBus.execute(command);
+
+    verify(handler).handle(command);
+  }
+
+
+}
